When posting a big invoice (>800 lines) for products that are valuated in real_time, the `_post` override in
`purchase_stock` can become a bottleneck for anglo-saxon companies. This method calls `_stock_account_anglo_saxon_reconcile_valuation` which performs reconciliations for all the invoice_lines.

Thanks to odoo/odoo#113680 it's now possible to batch the reconciliation process via the backend method `_reconcile_plan`. This gives various speedups by batching create, reducing the number of flush, etc.

Moving the reconciliation process outside of the loops in `_stock_account_anglo_saxon_reconcile_valuation` and calling `_reconcile_plan` only at the end of the method gives a significant speedup for big invoices.

speedup

Time to post an invoice changing the number of lines.

| # Lines |     Before PR            | After PR |
|:-------:|:------------------------:|:--------:|
|    10   |       1.66s              |   1.11s  |
|    50   |       8.53s              |   2.95s  |
|    200  |       1min30s            |   11.51s |
|    889  |       +20min (killed)    |   58.84s |

opw-3901468

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr