@robodoo 2024-05-23
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/166482).