@Aurelienvd 2024-06-03
@odoo/rd-logistics Hello 👋 Could someone have a look at this PR when they have the time ? Thank you !